Move Functions to Controller
Moved player "struct" to FinSimController along with the calcMoney functions. A getPlayerMoney function was also created to return to the attributes of the Player struct instance "player"
